+++ app/src/features/layout/layout.tsx
@@ -28,7 +28,7 @@ export function Layout() {
         <div
           className={cn(
             "bg-view-1 h-full overflow-x-hidden main-view transition-[margin,border] duration-150 border-border/25 ml-0 mb-1.5 mr-1.5 rounded-md rounded-br-sm border",
-            isSidebarCollapsed && "m-0 rounded-none border-transparent",
+            isSidebarCollapsed && "m-0 rounded-none border-0",
           )}
         >
           <SearchDialog />